+++ data_binding/lib/data_binding.dart
@@ -1,0 +1,4 @@
+library model_binding;
+
+export 'package:data_binding/widget/data_provider.dart';
+export 'package:data_binding/widget/data_inject.dart';