+++ model_binding/lib/model_binding.dart
@@ -1,4 +1,0 @@
-library model_binding;
-
-export 'package:model_binding/widget/model_provider.dart';
-export 'package:model_binding/widget/model_inject.dart';